+++ src/EVault/shared/types/LTVConfig.sol
@@ -8,11 +8,12 @@ import {LTVType} from "./LTVType.sol";
 import "../Constants.sol";
 
 struct LTVConfig {
-    uint40 targetTimestamp;
+    // Packed slot: 6 + 2 + 4 + 2 + 1 = 15
+    uint48 targetTimestamp;
     ConfigAmount targetLTV;
-    uint24 rampDuration;
+    uint32 rampDuration;
     ConfigAmount originalLTV;
-    bool initialized; // flag is se
+    bool initialized;
 }
 
 library LTVConfigLib {
@@ -38,8 +39,8 @@ library LTVConfigLib {
         return ConfigAmount.wrap(uint16(ltv));
     }
 
-    function setLTV(LTVConfig memory self, ConfigAmount targetLTV, uint24 rampDuration) internal view returns (LTVConfig memory newLTV) {
-        newLTV.targetTimestamp = uint40(block.timestamp + rampDuration);
+    function setLTV(LTVConfig memory self, ConfigAmount targetLTV, uint32 rampDuration) internal view returns (LTVConfig memory newLTV) {
+        newLTV.targetTimestamp = uint48(block.timestamp + rampDuration);
         newLTV.targetLTV = targetLTV;
         newLTV.rampDuration = rampDuration;
         newLTV.originalLTV = self.getLTV(LTVType.LIQUIDATION);
